Rename login input component
